Created Settings Page and Fixed Program.cs File
Created settings page, currently with 2 changeable settings, the name of the business and color schemes, a dark and light one.

Also fixed a bug with the Program.cs file in which it didn't pick up the methods/objects in the main class.

Also added a property file that gets loaded at the start of the program to load things like color scheme and business name, it can be changed on the settings page.

NB! Once you click on settings page it will stay on the settings page, this is normal, if there were other forms it would rather switch to them, will add a close button to the main form for the child forms soon.